Bug 34886: Restore 'Place reserve' prior to session
This is a pragmatic patch to restore the display of 'Place hold' prior
to patron login. When we have no Patron, we fall back to only checkout
get_onshelfholds_policy without a patron. When we have a patron, we use
IsAvailableForItemLevelRequest instead to get the more detailed
response.

This isn't perfect, but I can't see an easy way to restore this
functionality otherwise.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>