Bug 35104: Throw exception on store of invalid marcxml
This patch adds an exception Koha::Biblio::Metadata->store to prevent
erroneous characters in marcxml being stored into the database.

We also remove the cleanup of such characters from TransformHtmlToMarc
as introduced by bug 34549.

This should result in there being no way, outside of direct database
manipulation, of getting forbidden characters into our marcxml store.

We'll need to work through the codebase to catch the exception thrown to
give the end user the oportunities to act on the error and fix the data
at source.

Signed-off-by: Baptiste Bayche <baptiste.bayche@inlibro.com>
Signed-off-by: Baptiste Wojtkowski <baptiste.wojtkowski@biblibre.com>