Bug 35657: Add support for assignee_id to ticket_updates endpoint
This patch starts to add support for cross-synced ticket.assignee_id
updates.  You should be able to either set assignee at the time of
adding a ticket_update or directly on the ticket.. but in both cases end
up with a ticket_update with the fine details of when and who set the
assigee.