Bug 32565: Remove unallocated param from cron script
unallocated param is redundant, it is only relevant if RealTimeHoldsQueue is enabled, the cron should just fall back to unallocated mode if the sys pref is enabled.
This simplifies system administration, and does not require current Koha installations to update their cron configuration entry to make use of this fix.

I think we should add a documentation note recommending using this cronjob when RealTimeHoldsQueue is enabled