Bug 36285: [PoC] Move ERM specific styling to all of koha Staff forms
This is just a proof of concept

1) Before applying the patch, access some (or all) of the links below.
2) Apply patch
3) Run yarn css:build to rebuild scss
4) Refresh each of the pages and notice the changes

Some form examples:
<staff_url>/cgi-bin/koha/acqui/supplier.pl?op=enter
<staff_url>/cgi-bin/koha/serials/serials-home.pl
<staff_url>/cgi-bin/koha/serials/subscription-add.pl
<staff_url>/cgi-bin/koha/erm/eholdings/local/titles/add
<staff_url>/cgi-bin/koha/virtualshelves/shelves.pl?op=add_form&referer=list
<staff_url>/cgi-bin/koha/suggestion/suggestion.pl?op=add_form&branchcode=CPL

Side menu changes:
/cgi-bin/koha/members/members-home.pl
/cgi-bin/koha/serials/serials-search.pl # Press Search

A problem arises for some inputs that have the size property e.g. size="4" on the new purchase suggestion form, and others.
With this new change, these inputs lose their initial enforced "smaller" size.
I think this should be iterated and made more streamlined and use a .input-small class of the size.