Bug 36641: Add endpoint for fetching circulation rules
This patch is a proof of concept for an API endpoint that returns a list
of circulation rules for the specified itemtype, branchcode and patron
category combination.  You can also add a list of `rules` as a query
parameter to limit the response to only the rules you are interested in
for this combination.