Bug 33260: (QA follow-up) Safely continue for non-templated notices
The 'letter_id'/'letter_code' field in the message queue table is a
NULLable foreign key as is indeed NULL in certain circustances. We need
to gracefully degrade when there isn't a template for the notice in
question.

Tihs patch uses the newly introduced 'template' accessor and when that
returns undefined we pass 'style' as an empty string.

We also use the 'is_html' method introduced in bug 35285 to set the html
flag.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>