Bug 37301: [PoC] DRY ToolbarButton
This is only a PoC applied to Agreements for discussion.
This is a continuation of bug 37274.
Bug 37274 applied the ToolbarButton to missing components but it was mostly replacing WET for WET.

This is DRYer:
- All mark-up related to toolbar buttons is now written once (css class, icons, elements)
- All labels are now written once (except 'Add' for translatability)

Bug 37301: Remove methods

We do not need the methods, we can emit from the template.
Also rename the events to the wording more consistent with what we have.

Bug 37301: Rename 'type' with 'action'

'type' is too generic.
Also move to else-if and add a else block to ease debugging in case an
unsupported action is passed

Bug 37301: PoC (Continuation): Add BaseResource and AgreementResource

This is the building blocks for a much consistent approach, for example:
AgreementsShow -> delete_agreement && AgreementsList -> doDelete should be DRY into AgreementResource
Any WET code currently in setup for list+show of a resource can (and should) be moved into <resourceName>Resource
Any WET code currently in setup for all 'resources' can (and should) be moved into BaseResource

and others

Bug 37301: (follow-up) PoC (Continuation): Add BaseResource and AgreementResource

To squash with previous patch

Bug 37301: Remove toToResource* methods to BaseResource

Bug 37301: Add goToResourceShow to BaseResource

Bug 37301: Get rid of e.preventDefault

Feels wrong to have datatable+event related args in BaseResource, remove that.
The only reason we needed preventDefault was because 'show' action was a
link with an href, but we dont even need that, and add 'role="button"'
for cursor:pointer mouseover interaction.

Bug 37301: DRY delete

Bug 37301: Move table_url declaration to resource class

This is relevant because doResourceDelete expected table_url to be implemented, but LicensesList does not implement it for example.
We now have a getResourceTableUrl implemented by all resources

Bug 37301: Apply changes to licenses

Bug 37301: Apply changes to eHoldings local packages

Bug 37301: Restore possibility of a custom toolbar button

EHoldingsLocalTitles needs this for 'Import from list' and 'Import from KBART file'

Bug 37301: Apply changes to eholdings titles

Bug 37301: Move setWarning to BaseResource

UsageStatisticsDataProvidersList needs it, and others in the future will as well

Bug 37301: Fix: URL param needs to match attr_id

url param needs to match attr_id
usage statistics is non-standard when compared to agreements+licenses etc.
using usage_data_provider_id but resource id field is erm_usage_data_provider_id.
This patch is fixing that.

Bug 37301: Add required faicon for fontawesome compnent

Bug 37301: Apply changes to usage statistics data providers

This is also removing UsageStatissticsDataProvidersToolbar as that is
not required, and updating the UsageStatisticsDataProvidersSummary
button to a ToolbarButton instead of a router-link

Bug 37301: Apply changes to preservation trains

Also update edit route from TrainsFormEdit to TrainsFormAddEdit for consistency

Bug 37301: Update cypress

Local packages were named 'package' on the delete confirmation dialog, but 'local package' on the deleted confirmation message.
Both these places now use the same i18n.display_name from the respective Resource component so these label tests need to be updated.
Same for local titles.