Bug 37389: Fix performance problems with extended attributes on the API
This patch adds a number of methods that work together to result the
problem as a whole.

We add a builder method to dynamically create DBIx::Class relations when
required for the Koha::ILL::Requests and Koha::Patrons resultset.

We add a ExtendedAttributes.pm mixin to call the aforementioned builder
method so we can subsequently use these new relations to create a
performance join condition.

We that add an AUTOLOAD into Koha::ILL::Request to allow use of the
new dynamically added DBIx::Class relations in the above as
extended_attributes_$type from the corresponding Koha::Object.

Finally, we reworked the query builder code to ensure we're using the
new relations added when using the API.

Co-authored-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>