Bug 37389: [SQUASH] Join alias fix
This is to be squashed into the previous patch, but I'm leaving an explanation here of this follow-up fix for now.
Only rename the join aliases if extended_attribute entry is a string 'type'.
If its an array, skip.
This 'array' method is used on the left-side search forms for ILL and patrons table.
Its the difference between

      [
        {
          "extended_attributes.type": [
            "title",
            "type",
            "author",
            "article_title",
            "pages",
            "issue",
            "volume",
            "year"
          ],
          "extended_attributes.value": {
            "like": "%book%"
          }
        }
      ]

and

[
    {
      "extended_attributes.type": "year",
      "extended_attributes.value": {
        "like": "%book%"
      }
    }
  ],
  [
    {
      "extended_attributes.type": "pages",
      "extended_attributes.value": {
        "like": "%book%"
      }
    }
  ],

Where the former searches for a single keyword occurrence in a potential list of extended_attributes and the latter allows for different keywords for each extended_attribute type (even though the datatable search input allows only for 1, this is utilized in other places such as ERM licenses additional fields and subscriptions searchable additional fields I believe).