Bug 35044: (QA follow-up): Allow for value '0' in additional fields
This now allows for value '0' for an additional field (repeatable or not).
This still does not allow for an empty '' field in additional fields.
This is consistent with borrower patron attributes, i.e. if you submit an empty borrower patron attribute entry, it does not store it as empty.
Additionally, if you have a value in a borrower patron attribute but then submit the form with that field as empty, it gets deleted from the database.

I agree not being able to save the value '0' is a blocker, but not the empty.
For consistency sake I'm keeping this as is, I don't see a use-case where a row with an empty field value is preferred over it not existing.

If we allow for '' values of additional fields, the 'clear' button for repeatable fields would have to remove the input instead of clearing the value (or else how would the user remove repeatable fields entries?).
If this is done, it's inconsistent with the repeatable fields from patron attribute types, as in that instance the 'clear' button only removes the input contents, not the input element itself.

prove t/db_dependent/Koha/Object/Mixin/AdditionalFields.t